refactor: 서비스 역할 command, query로 분리
