+++ src/main/java/com/leets/commitatobe/domain/login/usecase/LoginCommandService.java
@@ -1,6 +1,7 @@
 package com.leets.commitatobe.domain.login.usecase;
 
 import com.leets.commitatobe.domain.login.presentation.dto.GitHubDto;
+import com.leets.commitatobe.domain.login.presentation.dto.JwtResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
@@ -8,10 +9,10 @@ public interface LoginCommandService {
     // 깃허브 accessToken가져오는 함수
     String gitHubLogin(String authCode);
 
-    // AUTH 헤더에서 엑세스 토큰을 이용해 유저 아이디를 불러오는 함수
-    GitHubDto getGitHubUser(HttpServletRequest request);
-
     // 로그인시 리다이렉트 처리 함수
     void redirect(HttpServletResponse response);
+
+    // 응답에 쿠키 설정 함수
+    void setRefreshTokenCookie(HttpServletResponse response, String refreshToken);
 
 }
