+++ src/main/java/com/leets/commitatobe/domain/login/usecase/LoginCommandServiceImpl.java
@@ -2,15 +2,13 @@ package com.leets.commitatobe.domain.login.usecase;
 
 import static com.leets.commitatobe.global.response.code.status.ErrorStatus._GITHUB_JSON_PARSING_ERROR;
 import static com.leets.commitatobe.global.response.code.status.ErrorStatus._GITHUB_TOKEN_GENERATION_ERROR;
-import static com.leets.commitatobe.global.response.code.status.ErrorStatus._JWT_NOT_FOUND;
 import static com.leets.commitatobe.global.response.code.status.ErrorStatus._REDIRECT_ERROR;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.leets.commitatobe.domain.login.domain.CustomUserDetails;
-import com.leets.commitatobe.domain.login.presentation.dto.GitHubDto;
+import com.leets.commitatobe.domain.user.domain.repository.UserRepository;
 import com.leets.commitatobe.global.exception.ApiException;
 import com.leets.commitatobe.global.utils.JwtProvider;
-import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.Collections;
 import java.util.Map;
@@ -22,7 +20,6 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 
@@ -41,6 +38,8 @@ public class LoginCommandServiceImpl implements LoginCommandService {
     private String redirectUri;
 
     private final JwtProvider jwtProvider;
+
+    private final UserRepository userRepository;
 
     // 예찬님! 이 함수 사용해서 accessToken 가져오심 됩니다~!
     public String gitHubLogin(String authCode) {
@@ -91,22 +90,6 @@ public class LoginCommandServiceImpl implements LoginCommandService {
     }
 
 
-    public GitHubDto getGitHubUser(HttpServletRequest request) {
-        String authorizationHeader = request.getHeader("Authorization");
-
-        if (authorizationHeader == null || !authorizationHeader.startsWith("Bearer ")) {
-            throw new ApiException(_JWT_NOT_FOUND);
-        }
-
-        String accessToken = authorizationHeader.substring(7); // "Bearer " 이후의 토큰 값만 추출
-        Authentication authentication = jwtProvider.getAuthentication(accessToken);
-
-        CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
-
-        return userDetails.getGitHubDto();
-    }
-
-
     public void redirect(HttpServletResponse response){
         // TODO: user로 scope를 설정하면 읽고 쓰는 권한 모두를 가져오게 됨, 이후 개발하며 개발 범위에 맞춰 수정이 필요함
         String url = "https://github.com/login/oauth/authorize?client_id=" + clientId + "&redirect_uri=" + redirectUri + "&scope=user";
@@ -116,6 +99,17 @@ public class LoginCommandServiceImpl implements LoginCommandService {
             throw new ApiException(_REDIRECT_ERROR);
         }
     }
+
+    @Override
+    public void setRefreshTokenCookie(HttpServletResponse response, String refreshToken) {
+        Cookie refreshTokenCookie = new Cookie("refreshToken", refreshToken);
+        refreshTokenCookie.setHttpOnly(true);
+        refreshTokenCookie.setSecure(true); // HTTPS를 사용할 경우
+        refreshTokenCookie.setPath("/");
+        refreshTokenCookie.setMaxAge(7 * 24 * 60 * 60); // 7일 동안 유효
+        response.addCookie(refreshTokenCookie);
+    }
+
 }
